Reorg code
 - Move selinuxrelabel to file.go
 - Move setupstyle to v1.Runconfig
