+++ pkg/utils/common.go
@@ -19,20 +19,12 @@ package utils
 import (
 	"fmt"
 	"github.com/rancher-sandbox/elemental-cli/pkg/types/v1"
-	"github.com/spf13/afero"
 	"io"
 	"os"
 	"os/exec"
 	"strings"
 )
 
-const (
-	GPT = "gpt"
-	ESP = "esp"
-	BIOS = "bios_grub"
-	MSDOS = "msdos"
-	BOOT = "boot"
-)
 
 func GetUrl(client v1.HTTPClient, url string, destination string) error {
 	var source io.Reader
@@ -63,46 +55,9 @@ func GetUrl(client v1.HTTPClient, url string, destination string) error {
 	return nil
 }
 
-func selinuxRelabel(target string, fs afero.Fs) error {
-	var err error
-
-	contextFile := fmt.Sprintf("%s/etc/selinux/targeted/contexts/files/file_contexts", target)
-
-	_, err1 := fs.Stat(contextFile)
-	contextExists := err1 == nil
-
-	if commandExists("setfiles") && contextExists {
-		_, err = exec.Command("setfiles", "-r", target, contextFile, target).CombinedOutput()
-	}
-
-	return err
-}
-
 func commandExists(command string) bool {
 	_, err := exec.LookPath(command)
 	return err == nil
-}
-
-func setupStyle(config *v1.RunConfig, fs afero.Fs) {
-	var part,boot string
-
-	_, err := fs.Stat("/sys/firmware/efi")
-	efiExists := err == nil
-
-
-	if config.ForceEfi || efiExists {
-		part = GPT
-		boot = ESP
-	} else if config.ForceGpt {
-		part = GPT
-		boot = BIOS
-	} else {
-		part = MSDOS
-		boot = BOOT
-	}
-
-	config.PartTable = part
-	config.BootFlag = boot
 }
 
 func BootedFrom(label string) bool {
