+++ pkg/utils/common_test.go
@@ -19,9 +19,7 @@ package utils
 import (
 	"fmt"
 	. "github.com/onsi/gomega"
-	v1 "github.com/rancher-sandbox/elemental-cli/pkg/types/v1"
 	"github.com/rancher-sandbox/elemental-cli/tests/mocks"
-	"github.com/spf13/afero"
 	"io/ioutil"
 	"os"
 	"testing"
@@ -60,7 +58,6 @@ func TestGetUrlNet(t *testing.T) {
 	Expect(f.Size()).To(Equal(int64(1024)))  // We create a 1024 bytes size file on the mock
 }
 
-
 func TestGetUrlFile(t *testing.T) {
 	RegisterTestingT(t)
 	client := &mocks.MockClient{}
@@ -97,36 +94,4 @@ func TestBootedFrom(t *testing.T) {
 	Expect(BootedFrom("")).To(BeTrue()) // Empty label should match everything!
 }
 
-func TestSetupStyleDefault(t *testing.T) {
-	RegisterTestingT(t)
-	fs := afero.NewMemMapFs()
-	c := v1.RunConfig{}
-	setupStyle(&c, fs)
-	Expect(c.PartTable).To(Equal(MSDOS))
-	Expect(c.BootFlag).To(Equal(BOOT))
-	c = v1.RunConfig{
-		ForceEfi: true,
-	}
-	setupStyle(&c, fs)
-	Expect(c.PartTable).To(Equal(GPT))
-	Expect(c.BootFlag).To(Equal(ESP))
-	c = v1.RunConfig{
-		ForceGpt: true,
-	}
-	setupStyle(&c, fs)
-	Expect(c.PartTable).To(Equal(GPT))
-	Expect(c.BootFlag).To(Equal(BIOS))
-}
 
-func TestSelinuxRelabel(t *testing.T) {
-	// This testing is ridiculous as this function can fail or not and we dont really care, we cannot return the proper error
-	// So we are not really testing
-	// Also, I cant seem to mock  exec.LookPath so it will always fail tor un due setfiles not being in the system :/
-	RegisterTestingT(t)
-	fs := afero.NewMemMapFs()
-	// This is actually failing but not sure we should return an error
-	Expect(selinuxRelabel("/", fs)).To(BeNil())
-	fs = afero.NewMemMapFs()
-	_, _ = fs.Create("/etc/selinux/targeted/contexts/files/file_contexts")
-	Expect(selinuxRelabel("/", fs)).To(BeNil())
-}
